Use CMake install dirs
Currently we use KDEInstallDirs from ECM. This means adds a dependency on ECM that we otherwise don't need.

Futhermore KDEInstallDirs internally relies on Qt to determine some install locations, even though that's not needed here. This adds additional complexity to the build.

To avoid this use the values from CMake's GNUInstallDirs